requirements.txt
PyYAML
flask
requests
gevent
pillow
Keras
tensorflow
keras-tuner